requirements.txt
annotated-types==0.7.0
bitarray==2.9.2
bitstring==4.2.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.70
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
ecdsa==0.19.0
esp-coredump==1.12.0
esp-idf-kconfig==2.3.0
esp-idf-monitor==1.5.0
esp-idf-panic-decoder==1.2.1
esp-idf-size==1.6.0
esp_idf_nvs_partition_gen==0.1.3
esptool==4.8.1
fonttools==4.55.0
freertos-gdb==1.0.3
idf-component-manager==2.0.4
idna==3.10
intelhex==2.3.0
jsonref==1.1.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.3
packaging==24.1
pillow==11.0.0
pyclang==0.5.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyelftools==0.31
pygdbmi==0.11.0.0
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.2
six==1.16.0
tk==0.1.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
windows-curses==2.3.3
